% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpercluster.R, R/oxy-countpercluster.R
\name{countpercluster}
\alias{countpercluster}
\title{count cell numbers per each cluster for each group}
\usage{
countpercluster(data, group, level)

countpercluster(data, group, level)
}
\arguments{
\item{data}{A Seurat object}

\item{group}{the name of metadata column within the Seurat object that indicate groups you want to make comparison}

\item{level}{the order of group names you want in your plot}
}
\value{
A table containing cell numbers in each cluster for each group. Bar charts showing percentage of cell type in each group.

A table containing cell numbers in each cluster for each group. Bar charts showing percentage of cell type in each group.
}
\description{
This function quickly calculates and visualizes cell numbers and their percentages in each cluster for each group.

This function quickly calculates and visualizes cell numbers and their percentages in each cluster for each group.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 countpercluster(skin_tcell,group = "tissue",level=unique(skin_tcell@meta.data$tissue))
 }
}
}
