% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_heatmap.R,
%   R/oxy-annotation_heatmap.R
\name{annotation_heatmap}
\alias{annotation_heatmap}
\title{count cell numbers in each group in each cluster}
\usage{
annotation_heatmap(data, gene, group)

annotation_heatmap(data, gene, group)
}
\arguments{
\item{data}{A Seurat object}

\item{gene}{a list of genes for heatmap}

\item{group}{the name of meta.data column that you want to split the dataset for comparison}
}
\value{
A heatmap plot showing mean expression of each gene in each group or percentage of cells that have no-zero expression of each gene in each group

A heatmap plot showing mean expression of each gene in each group or percentage of cells that have no-zero expression of each gene in each group
}
\description{
This function applies to 10X Visium data and allows you to check cell numbers in each group in each cluster.

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
