% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_heatmap.R,
%   R/oxy-annotation_heatmap.R
\name{annotation_heatmap}
\alias{annotation_heatmap}
\title{plot heatmap to visualize expression of provided key markers in each cell cluster or experimental group}
\usage{
annotation_heatmap(data, gene, group)
}
\arguments{
\item{data}{A Seurat object}

\item{gene}{a list of genes for heatmap eg. gene<-c("Cd44","Cd69","Itgae","Ccr7","Il7r","Tbx21","Ifng","Il2","Gata3","Il4","Il5","Foxp3","Ctla4","Pdcd1","Tigit","Il2ra","Gzmb","Il10","Il17a","Cxcr5","Bcl6","Cd8a","Cd4")}

\item{group}{the name of meta.data column within the Seurat object that indicates the cell cluster eg. group<-"seurat_clusters"}
}
\value{
A heatmap plot showing mean expression or percentage of cells that have no-zero expression of each gene in each cell cluster
}
\description{
plot heatmap to visualize expression of provided key markers in each cell cluster or experimental group
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 annotation_heatmap(skin_tcell,gene = c("Cd44","Cd69"), group = "seurat_clusters")
 }
}
}
