% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpercluster.R, R/oxy-countpercluster.R
\name{countpercluster}
\alias{countpercluster}
\title{count cell numbers in each group in each cluster}
\usage{
countpercluster(data, group, plot, plot.cluster)

countpercluster(data, group, plot, plot.cluster)
}
\arguments{
\item{data}{A Seurat object}

\item{group}{the name of meta.data column that you want to split the dataset for comparison}

\item{plot}{whether to plot mean expression and percentage of total in each group for each cluster. plot="TRUE"/"T" otherwise no plot}

\item{plot.cluster}{select cluster to be included in the plot eg. plot.cluster="C0|C8"}
}
\value{
A table contains cell numbers in each group in each cluster. Plots showing mean expression and percentage of total in each group for each cluster

A table contains cell numbers in each group in each cluster. Plots showing mean expression and percentage of total in each group for each cluster
}
\description{
This function applies to 10X Visium data and allows you to check cell numbers in each group in each cluster.

This function applies to 10X scRNA-seq data and allows you to check cell numbers in each group in each cluster and make a comparison plot.
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
