% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countpercluster.R, R/oxy-countpercluster.R
\name{countpercluster}
\alias{countpercluster}
\title{count cell numbers in each group in each cluster}
\usage{
countpercluster(data, group, plot, plot.cluster)

countpercluster(data, group, plot, plot.cluster)
}
\arguments{
\item{data}{PARAM_DESCRIPTION}

\item{group}{PARAM_DESCRIPTION}

\item{plot}{PARAM_DESCRIPTION}

\item{plot.cluster}{PARAM_DESCRIPTION}
}
\value{
A table contains cell numbers in each group in each cluster

OUTPUT_DESCRIPTION
}
\description{
This function applies to 10X Visium data and allows you to check cell numbers in each group in each cluster.

FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
