% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/czplots-package.R
\docType{package}
\name{czplots-package}
\alias{czplots}
\alias{czplots-package}
\title{czplots: Wrapped Functions to Provide Fast Visulization Appraches}
\description{
Wrapped Functions to Provide Fast Visulization Appraches.
}
\author{
\strong{Maintainer}: Chenhao Zhou \email{chenhao.zhou@uq.edu.au} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
